requirements-slim.txt
# Direct dependencies of the project. The full frozen pin list lives in
# `requirements.txt` (used by colab/setup.py). This file is the minimal
# set of packages a contributor needs to read or run the iteration code.

# Core scientific stack
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-image>=0.21
scikit-learn>=1.3
matplotlib>=3.7
Pillow>=10.0
tqdm>=4.65

# Deep learning
torch>=2.0
torchvision>=0.15

# Detection (Grounding DINO via OpenMMLab)
mmcv>=2.0
mmdet>=3.0
mmengine>=0.8
openmim>=0.3
transformers>=4.30
tokenizers
pycocotools

# Notebook tooling
jupyterlab
ipywidgets
nbformat
nbconvert

# MedSAM - installed editable from MedSAM/ directory:
#   pip install -e MedSAM/
# (kept out of this requirements file so that a `pip install -r` does not
# try to clone an unrelated git URL.)
